Webpack Update
Updating webpack to newest version for better compatibility with the
other dependencies. Unfortunately, the combination of wasm, web workers
and TypeScript seems to be flakey. Following the example on
https://webpack.js.org/loaders/worker-loader to use a config collides
with using 'web-worker!' url. Using import fails due to type and export
problems. The logic for wasm loading has to be disabled, because
emscripten already creates a js wrapper. Using import instead of quire
in the worker as recommended by webpack leads to use of the undefined
window variable in the worker.
The current state works, for now. No guarantees that anything keeps working
if further upgrades are performed.