Selection / Menu update redesign
The Menu is now no longer cached in a member, but via memoization. This
is simpler, since the property is now the "source of truth" for the menu
data. Similarly, the selection is now either read from the props or state
field, depending on whether Jeri is controlled or stand-alone. In controlled
mode, state is not used at all, and selection is always used from props.
Memoization should make this more efficient.